+++ services/web/server/src/simcore_service_webserver/activity/module_setup.py
@@ -9,7 +9,7 @@ from servicelib.aiohttp.rest_routing import (
 
 from .._constants import APP_OPENAPI_SPECS_KEY
 from . import handlers
-from .config import assert_valid_config
+from .settings import assert_valid_config
 
 logger = logging.getLogger(__name__)
